# api/api-spec.yaml
openapi: 3.0.0
servers:
  - description: Local Host
    url: http://localhost:5000
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/thomasyuan/TradeRev/1.0
info:
  version: "1.0.0"
  title: Acution API
  contact:
    email: thomas@cppcoder.net
...
paths:
  /accounts:
    post:
      tags:
        - Account
      summary: create a new account
      description: |
        Create a new account.
      requestBody:
        description: account information
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Account"
      responses:
        "200":
          description: account is created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JWT"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /me:
    get:
      tags:
        - Account
      security:
        - BearerAuth: []
      summary: get current user information
      description: |
        Get current user information
      responses:
        "200":
          description: fetch account information success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountGet"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      tags:
        - Account
      summary: update account information
      description: |
        Update account information. After get the account id from flinks, client should call this endpoint to update account info, and get a JSON Web Token, which should be sent with the subsequent request.
      security:
        - BearerAuth: []
      requestBody:
        description: account information
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountPatch"
      responses:
        "200":
          description: account is updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Account:
      type: object
      required:
        - id
        - name
        - email
        - password
      properties:
        id:
          type: string
          example: johnsmith
        name:
          type: string
          example: John Smith
        email:
          type: string
          example: "John.Smith@gmail.com"
        password:
          type: string
          example: "secrect"
    AccountGet:
      type: object
      required:
        - id
        - name
        - email
      properties:
        id:
          type: string
          example: johnsmith
        name:
          type: string
          example: John Smith
        email:
          type: string
          example: "John.Smith@gmail.com"
    AccountPatch:
      type: object
      properties:
        name:
          type: string
          example: John Smith
        email:
          type: string
          example: "John.Smith@gmail.com"
        password:
          type: string
          example: "secrect"
...
    JWT:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
  responses:
    BadRequest:
      description: The request is incorrect or corrupted
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: number
                format: int32
                example: 400
              message:
                type: string
                example: "The request is incorrect or corrupted"
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: number
                format: int32
                example: 404
              message:
                type: string
                example: "The specified resource was not found"
    Unauthorized:
      description: The client is unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: number
                format: int32
                example: 401
              message:
                type: string
                example: "The client is unauthorized"
    Forbidden:
      description: The client doesn't not have permission
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: number
                format: int32
                example: 403
              message:
                type: string
                example: "The client does not have permission"